Create README.md
After Remy returns to help Linguini, Linguini finally reveals to the rest of the kitchen staff the truth that Remy is responsible for the restaurant’s renewed success. Everyone quits. Remy’s family rallies to help him and Linguini serve the guests—and Anton. Colette returns and helps Remy create a special dish of ratatouille for Anton, who loves it and demands to meet the cook.